Minor fix - inversione codice sqlEditor <-> app
