+++ sqlEditor.node.js
@@ -1,4 +1,4 @@
-//Connessione SQL
+//Semplice editor Sql in node
 var mysql      = require('mysql');
 var connection = mysql.createConnection({
   host     : 'ritardoautobus.c147tajn45vc.us-east-2.rds.amazonaws.com',
@@ -6,11 +6,7 @@ var connection = mysql.createConnection({
   password : 'trentino',
   database : 'ritardoautobus'
 });
-//Instanza Express
-var express = require('express');
-var app=express();
 
-/*
 //Query
 connection.connect();
 
@@ -22,14 +18,3 @@ connection.query('SELECT * from test', function(err, rows, fields) {
 });
 
 connection.end();
-*/
-var segnaRitardo = function(params){
-  var query = ('INSERT into qualcosa');
-  connection.connect();
-  connection.Query(query,function (errore){
-    if(!errore)
-      console.log('Aggiunta riga ai ritardi');
-    else
-      console.log('Errore nell\'aggiunta dei ritardi');
-  })
-}
